% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_vv_cos}
\alias{fast_vv_cos}
\title{vForce cosine (Tier 2c)}
\usage{
fast_vv_cos(v)
}
\arguments{
\item{v}{Numeric vector}
}
\value{
Numeric vector with cos(v) elementwise via Apple Accelerate vvcos
}
\description{
vForce cosine (Tier 2c)
}
