% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_vv_exp}
\alias{fast_vv_exp}
\title{vForce exponential (Tier 2c)}
\usage{
fast_vv_exp(v)
}
\arguments{
\item{v}{Numeric vector}
}
\value{
Numeric vector with exp(v) elementwise via Apple Accelerate vvexp
}
\description{
vForce exponential (Tier 2c)
}
