% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_vv_log}
\alias{fast_vv_log}
\title{vForce logarithm (Tier 2c)}
\usage{
fast_vv_log(v)
}
\arguments{
\item{v}{Numeric vector}
}
\value{
Numeric vector with log(v) elementwise via Apple Accelerate vvlog
}
\description{
vForce logarithm (Tier 2c)
}
