% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_vv_sin}
\alias{fast_vv_sin}
\title{vForce sine (Tier 2c)}
\usage{
fast_vv_sin(v)
}
\arguments{
\item{v}{Numeric vector}
}
\value{
Numeric vector with sin(v) elementwise via Apple Accelerate vvsin
}
\description{
vForce sine (Tier 2c)
}
